# MaritimumClient/Model/Port.cs
/*
 * Maritimum
 *
 * a sample API to return ships and ports information.
 *
 * The version of the OpenAPI document: v1
 * Generated by: https://github.com/openapitools/openapi-generator.git
 */


using System;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.Linq;
using System.IO;
using System.Runtime.Serialization;
using System.Text;
using System.Text.RegularExpressions;
using Newtonsoft.Json;
using Newtonsoft.Json.Converters;
using Newtonsoft.Json.Linq;
using System.ComponentModel.DataAnnotations;
using OpenAPIDateConverter = MaritimumClient.Client.OpenAPIDateConverter;

namespace MaritimumClient.Model
{
    /// <summary>
    /// Port
    /// </summary>
    [DataContract(Name = "Port")]
    public partial class Port : IEquatable<Port>, IValidatableObject
    {
        /// <summary>
        /// Initializes a new instance of the <see cref="Port" /> class.
        /// </summary>
        /// <param name="uuid">UUID of the port..</param>
        /// <param name="name">Name of the port..</param>
        /// <param name="country">Port&#39;s country..</param>
        /// <param name="isDeepWater">Whether this is a deep water port..</param>
        public Port(Guid uuid = default(Guid), string name = default(string), string country = default(string), bool? isDeepWater = default(bool?))
        {
            this.Uuid = uuid;
            this.Name = name;
            this.Country = country;
            this.IsDeepWater = isDeepWater;
        }

        /// <summary>
        /// UUID of the port.
        /// </summary>
        /// <value>UUID of the port.</value>
        [DataMember(Name = "uuid", EmitDefaultValue = false)]
        public Guid Uuid { get; set; }

        /// <summary>
        /// Name of the port.
        /// </summary>
        /// <value>Name of the port.</value>
        [DataMember(Name = "name", EmitDefaultValue = false)]
        public string Name { get; set; }

        /// <summary>
        /// Port&#39;s country.
        /// </summary>
        /// <value>Port&#39;s country.</value>
        [DataMember(Name = "country", EmitDefaultValue = false)]
        public string Country { get; set; }

        /// <summary>
        /// Whether this is a deep water port.
        /// </summary>
        /// <value>Whether this is a deep water port.</value>
        [DataMember(Name = "is_deep_water", EmitDefaultValue = true)]
        public bool? IsDeepWater { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class Port {\n");
            sb.Append("  Uuid: ").Append(Uuid).Append("\n");
            sb.Append("  Name: ").Append(Name).Append("\n");
            sb.Append("  Country: ").Append(Country).Append("\n");
            sb.Append("  IsDeepWater: ").Append(IsDeepWater).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public virtual string ToJson()
        {
            return Newtonsoft.Json.JsonConvert.SerializeObject(this, Newtonsoft.Json.Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="input">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object input)
        {
            return this.Equals(input as Port);
        }

        /// <summary>
        /// Returns true if Port instances are equal
        /// </summary>
        /// <param name="input">Instance of Port to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(Port input)
        {
            if (input == null)
                return false;

            return 
                (
                    this.Uuid == input.Uuid ||
                    (this.Uuid != null &&
                    this.Uuid.Equals(input.Uuid))
                ) && 
                (
                    this.Name == input.Name ||
                    (this.Name != null &&
                    this.Name.Equals(input.Name))
                ) && 
                (
                    this.Country == input.Country ||
                    (this.Country != null &&
                    this.Country.Equals(input.Country))
                ) && 
                (
                    this.IsDeepWater == input.IsDeepWater ||
                    (this.IsDeepWater != null &&
                    this.IsDeepWater.Equals(input.IsDeepWater))
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                int hashCode = 41;
                if (this.Uuid != null)
                    hashCode = hashCode * 59 + this.Uuid.GetHashCode();
                if (this.Name != null)
                    hashCode = hashCode * 59 + this.Name.GetHashCode();
                if (this.Country != null)
                    hashCode = hashCode * 59 + this.Country.GetHashCode();
                if (this.IsDeepWater != null)
                    hashCode = hashCode * 59 + this.IsDeepWater.GetHashCode();
                return hashCode;
            }
        }

        /// <summary>
        /// To validate all properties of the instance
        /// </summary>
        /// <param name="validationContext">Validation context</param>
        /// <returns>Validation Result</returns>
        public IEnumerable<System.ComponentModel.DataAnnotations.ValidationResult> Validate(ValidationContext validationContext)
        {
            yield break;
        }
    }

}