# MaritimumClient/Model/Ship.cs
/*
 * Maritimum
 *
 * a sample API to return ships and ports information.
 *
 * The version of the OpenAPI document: v1
 * Generated by: https://github.com/openapitools/openapi-generator.git
 */


using System;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.Linq;
using System.IO;
using System.Runtime.Serialization;
using System.Text;
using System.Text.RegularExpressions;
using Newtonsoft.Json;
using Newtonsoft.Json.Converters;
using Newtonsoft.Json.Linq;
using System.ComponentModel.DataAnnotations;
using OpenAPIDateConverter = MaritimumClient.Client.OpenAPIDateConverter;

namespace MaritimumClient.Model
{
    /// <summary>
    /// Ship
    /// </summary>
    [DataContract(Name = "Ship")]
    public partial class Ship : IEquatable<Ship>, IValidatableObject
    {
        /// <summary>
        /// Initializes a new instance of the <see cref="Ship" /> class.
        /// </summary>
        /// <param name="uuid">UUID of the ship..</param>
        /// <param name="name">Name of the ship..</param>
        /// <param name="homePortUuid">UUID of home port..</param>
        /// <param name="tonnage">Age of the person..</param>
        public Ship(Guid uuid = default(Guid), string name = default(string), Guid homePortUuid = default(Guid), int? tonnage = default(int?))
        {
            this.Uuid = uuid;
            this.Name = name;
            this.HomePortUuid = homePortUuid;
            this.Tonnage = tonnage;
        }

        /// <summary>
        /// UUID of the ship.
        /// </summary>
        /// <value>UUID of the ship.</value>
        [DataMember(Name = "uuid", EmitDefaultValue = false)]
        public Guid Uuid { get; set; }

        /// <summary>
        /// Name of the ship.
        /// </summary>
        /// <value>Name of the ship.</value>
        [DataMember(Name = "name", EmitDefaultValue = false)]
        public string Name { get; set; }

        /// <summary>
        /// UUID of home port.
        /// </summary>
        /// <value>UUID of home port.</value>
        [DataMember(Name = "home_port_uuid", EmitDefaultValue = false)]
        public Guid HomePortUuid { get; set; }

        /// <summary>
        /// Age of the person.
        /// </summary>
        /// <value>Age of the person.</value>
        [DataMember(Name = "tonnage", EmitDefaultValue = true)]
        public int? Tonnage { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class Ship {\n");
            sb.Append("  Uuid: ").Append(Uuid).Append("\n");
            sb.Append("  Name: ").Append(Name).Append("\n");
            sb.Append("  HomePortUuid: ").Append(HomePortUuid).Append("\n");
            sb.Append("  Tonnage: ").Append(Tonnage).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public virtual string ToJson()
        {
            return Newtonsoft.Json.JsonConvert.SerializeObject(this, Newtonsoft.Json.Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="input">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object input)
        {
            return this.Equals(input as Ship);
        }

        /// <summary>
        /// Returns true if Ship instances are equal
        /// </summary>
        /// <param name="input">Instance of Ship to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(Ship input)
        {
            if (input == null)
                return false;

            return 
                (
                    this.Uuid == input.Uuid ||
                    (this.Uuid != null &&
                    this.Uuid.Equals(input.Uuid))
                ) && 
                (
                    this.Name == input.Name ||
                    (this.Name != null &&
                    this.Name.Equals(input.Name))
                ) && 
                (
                    this.HomePortUuid == input.HomePortUuid ||
                    (this.HomePortUuid != null &&
                    this.HomePortUuid.Equals(input.HomePortUuid))
                ) && 
                (
                    this.Tonnage == input.Tonnage ||
                    (this.Tonnage != null &&
                    this.Tonnage.Equals(input.Tonnage))
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                int hashCode = 41;
                if (this.Uuid != null)
                    hashCode = hashCode * 59 + this.Uuid.GetHashCode();
                if (this.Name != null)
                    hashCode = hashCode * 59 + this.Name.GetHashCode();
                if (this.HomePortUuid != null)
                    hashCode = hashCode * 59 + this.HomePortUuid.GetHashCode();
                if (this.Tonnage != null)
                    hashCode = hashCode * 59 + this.Tonnage.GetHashCode();
                return hashCode;
            }
        }

        /// <summary>
        /// To validate all properties of the instance
        /// </summary>
        /// <param name="validationContext">Validation context</param>
        /// <returns>Validation Result</returns>
        public IEnumerable<System.ComponentModel.DataAnnotations.ValidationResult> Validate(ValidationContext validationContext)
        {
            yield break;
        }
    }

}